feat: implement matching restAPI for GET connecting to mongoDB
